+++ src/main/java/pro/sky/javacoursepart2/hw29/model/Employee.java
@@ -54,10 +54,6 @@ public class Employee {
         return lastName;
     }
 
-    public static int getCounter() {
-        return counter;
-    }
-
     public int getId() {
         return id;
     }
